+++ test/dummy/app/concepts/todos/operation/update.rb
@@ -1,10 +1,13 @@
 module Todos::Operation
   class Update < Trailblazer::Operation
-    step ->(options, params) { options["representer.render.class"] = Todos::Representer::Create }
-    step ->(options, params) { options["representer.errors.class"] = Reativo::Representer::Errors }
-    step Model( Todo , :find_cuid )
+    class Present < Trailblazer::Operation
+      step ->(options, params) { options["representer.render.class"] = Todos::Representer::Create }
+      step ->(options, params) { options["representer.errors.class"] = Reativo::Representer::Errors }
+      step Model( Todo , :find_cuid )
+      step Contract::Build(constant: Todos::Contract::Update)
+    end
 
-    step Contract::Build(constant: Todos::Contract::Update)
+    step Nested(Present)
     step Contract::Validate(key: 'todo')
     step Contract::Persist()
   end