+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - [JS] React Select with Auto Complete
 
+### Fixed
+- Crud Controller calls right operations new (Create::Present), and edit (Update::Present)
+
 ## [0.1.3] - 2019-03-26
 ### Added
 - [JS] Add babel-polyfill