+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Crud Controller calls right operations new (Create::Present), and edit (Update::Present)
+- Template for operation Index dont call Model.all to avoid data leaking
 
 ## [0.1.3] - 2019-03-26
 ### Added