+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.1.4] - 2019-03-28
 ### Added
 - [JS] React Select with Auto Complete
 - Crud Controller merges current_user into trailblazer run options
@@ -42,7 +44,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Generator
 - Reativo Crud Controller
 
-[Unreleased]: https://github.com/fernandes/reativo/compare/v0.1.3...HEAD
+[Unreleased]: https://github.com/fernandes/reativo/compare/v0.1.4...HEAD
+[0.1.4]: https://github.com/fernandes/reativo/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/fernandes/reativo/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/fernandes/reativo/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/fernandes/reativo/compare/v0.1.0...v0.1.1